DeltaTime calculations

+++ ProceduralWorldGenerator/src/Core/Application.h
@@ -27,7 +27,7 @@ private:
 
 	bool m_Running = true;
 	bool m_Minimized = false;
-	float m_LastFrameTime = 0.0f;
+	double m_LastFrameTime = 0.0;
 };
